+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## next
 
+## `@svizzle/barchart` v0.5.1 (next)
+
+- use reference values to build the x scale in order to draw ref lines exceeding the data extent correctly
+
 ## `@svizzle/choropleth` v0.6.0 (next)
 
 - add prop `projection`